+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo.hs
@@ -1,13 +1,9 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
--- CanStartFromGenesis
-{-# OPTIONS_GHC -Wno-deprecations #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Alonzo (
@@ -23,7 +19,6 @@ module Cardano.Ledger.Alonzo (
 where
 
 import Cardano.Ledger.Alonzo.Era
-import Cardano.Ledger.Alonzo.Genesis
 import Cardano.Ledger.Alonzo.PParams ()
 import Cardano.Ledger.Alonzo.Plutus.TxInfo ()
 import Cardano.Ledger.Alonzo.Rules ()
@@ -35,15 +30,13 @@ import Cardano.Ledger.Alonzo.TxAuxData (AlonzoTxAuxData)
 import Cardano.Ledger.Alonzo.TxBody (AlonzoTxBody, AlonzoTxOut)
 import Cardano.Ledger.Alonzo.TxWits ()
 import Cardano.Ledger.Alonzo.UTxO ()
-import Cardano.Ledger.BaseTypes (Globals)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Crypto (Crypto, StandardCrypto)
-import Cardano.Ledger.Keys (DSignable, Hash)
+import Cardano.Ledger.Keys (DSignable)
 import Cardano.Ledger.Mary.Value (MaryValue)
 import Cardano.Ledger.Plutus.Data ()
 import Cardano.Ledger.Rules.ValidationMode (applySTSNonStatic)
-import qualified Cardano.Ledger.Shelley.API as API
-import Cardano.Ledger.Shelley.API.Mempool
+import Cardano.Ledger.Shelley.API
 import Control.Arrow (left)
 import Control.Monad.Except (MonadError, liftEither)
 import Control.Monad.Reader (runReader)
@@ -55,7 +48,7 @@ type Alonzo = AlonzoEra StandardCrypto
 
 reapplyAlonzoTx ::
   forall era m.
-  (API.ApplyTx era, MonadError (ApplyTxError era) m) =>
+  (ApplyTx era, MonadError (ApplyTxError era) m) =>
   Globals ->
   MempoolEnv era ->
   MempoolState era ->
@@ -66,14 +59,10 @@ reapplyAlonzoTx globals env state vtx =
         flip runReader globals
           . applySTSNonStatic
             @(EraRule "LEDGER" era)
-          $ TRC (env, state, API.extractTx vtx)
-   in liftEither . left API.ApplyTxError $ res
+          $ TRC (env, state, extractTx vtx)
+   in liftEither . left ApplyTxError $ res
 
-instance (Crypto c, DSignable c (Hash c EraIndependentTxBody)) => API.ApplyTx (AlonzoEra c) where
+instance (Crypto c, DSignable c (Hash c EraIndependentTxBody)) => ApplyTx (AlonzoEra c) where
   reapplyTx = reapplyAlonzoTx
 
-instance (Crypto c, DSignable c (Hash c EraIndependentTxBody)) => API.ApplyBlock (AlonzoEra c)
-
-instance Crypto c => API.CanStartFromGenesis (AlonzoEra c) where
-  type AdditionalGenesisConfig (AlonzoEra c) = AlonzoGenesis
-  fromShelleyPParams ag = translateEra' ag . API.fromShelleyPParams ()
+instance (Crypto c, DSignable c (Hash c EraIndependentTxBody)) => ApplyBlock (AlonzoEra c)