+++ eras/conway/impl/src/Cardano/Ledger/Conway.hs
@@ -1,13 +1,7 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
--- CanStartFromGenesis
-{-# OPTIONS_GHC -Wno-deprecations #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Conway (
@@ -20,7 +14,6 @@ import Cardano.Ledger.Alonzo (reapplyAlonzoTx)
 import Cardano.Ledger.Babbage.TxBody ()
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Era (ConwayEra)
-import Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
 import Cardano.Ledger.Conway.Governance (RunConwayRatify (..))
 import Cardano.Ledger.Conway.Rules ()
 import Cardano.Ledger.Conway.Transition ()
@@ -30,30 +23,16 @@ import Cardano.Ledger.Conway.TxInfo ()
 import Cardano.Ledger.Conway.TxOut ()
 import Cardano.Ledger.Conway.UTxO ()
 import Cardano.Ledger.Crypto (Crypto, StandardCrypto)
-import Cardano.Ledger.Keys (DSignable, Hash)
-import qualified Cardano.Ledger.Shelley.API as API
+import Cardano.Ledger.Keys (DSignable)
+import Cardano.Ledger.Shelley.API
 
 type Conway = ConwayEra StandardCrypto
 
 -- =====================================================
 
-instance
-  ( Crypto c
-  , DSignable c (Hash c EraIndependentTxBody)
-  ) =>
-  API.ApplyTx (ConwayEra c)
-  where
+instance (Crypto c, DSignable c (Hash c EraIndependentTxBody)) => ApplyTx (ConwayEra c) where
   reapplyTx = reapplyAlonzoTx
 
-instance
-  ( Crypto c
-  , DSignable c (Hash c EraIndependentTxBody)
-  ) =>
-  API.ApplyBlock (ConwayEra c)
-
-instance Crypto c => API.CanStartFromGenesis (ConwayEra c) where
-  type AdditionalGenesisConfig (ConwayEra c) = ConwayGenesis c
-  fromShelleyPParams =
-    error "Unimplemented: Current interface is too limited and needs replacement for Conway to work"
+instance (Crypto c, DSignable c (Hash c EraIndependentTxBody)) => ApplyBlock (ConwayEra c)
 
 instance Crypto c => RunConwayRatify (ConwayEra c)